+++ src/just_makeit/_glue.py
@@ -171,6 +171,7 @@ def component_ctx(
             ),
             enums=C.enums(cfg),  # gh-1021
             records=C.records(cfg, object_name),  # gh-1405
+            properties=C.properties(cfg, object_name),  # gh-1426
         )
     )
     for _slot in _override_slots: